+++ src/dashboards/dto/index.ts
@@ -1,6 +1,7 @@
 export * from './current-month-by-service-query.dto';
 export * from './current-month-by-service-response.dto';
 export * from './dashboard-query.dto';
+export * from './inspectors-production-response.dto';
 export * from './low-score-collaborators-query.dto';
 export * from './low-score-collaborators-response.dto';
 export * from './non-conformities-by-checklist-query.dto';